requirements.txt
aioboto3==12.3.0
aiobotocore==2.11.2
aiohttp==3.9.3
aiohttp-cors==0.7.0
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
blessed==1.20.0
boto3==1.34.34
botocore==1.34.34
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.6
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
google-api-core==2.17.1
google-auth==2.28.2
googleapis-common-protos==1.62.0
gpustat==1.1.1
grpcio==1.62.1
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
py-spy==0.3.14
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ray==2.9.3
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smart-open==7.0.1
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.24
urllib3==2.0.7
virtualenv==20.25.1
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
